change the macro name LENDIAN to REVERSE_ENDIAN
defining this macro indicates that the endianness should be reversed.
when the host machine is little-endian architecture and the endianness of the target MIPS binary code is big-endian, this macro should be defined.
when the endiannesses of the host machine and the target MIPS binary is same, this macro should NOT be defined.

NOTE:
Now, it is assumed that the host machine is little-endian such as x86 and ARM, and that the target MIPS binary code is also little-endian.
(the original SESC simulator assumed MIPSEB (big-endian MIPS) binary codes.)